Trait sql_query_connector::FromSource

pub trait FromSource {
    fn from_source<'life0, 'life1, 'async_trait>(
        source: &'life0 Datasource, 
        url: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Self>> + Send + 'async_trait>>
    where
        Self: Connector + Sized,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Required methods

fn from_source<'life0, 'life1, 'async_trait>(
source: &'life0 Datasource,
url: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Self>> + Send + 'async_trait>> where
Self: Connector + Sized,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Instantiate a query connector from a Datasource.

The resolved url is passed distinctly from the datasource for two reasons:

Extracting the final url from the datasource involves resolving env vars and validating, which can fail with a schema parser error. We want to handle this as early as possible and in a single place.

The url may be modified with the config dir, in the case of Node-API.